very hacky tear down and build up of db, we'll see if this works for GitHub workflow to make tests work

+++ scripts/db-uninstall.sh
@@ -1,3 +1,4 @@
-docker stop timescaledb
-docker rm timescaledb
-docker rmi timescale/timescaledb:latest-pg14
+if [ "$(docker ps -q -f name=timescaledb)" ]; then
+    docker stop timescaledb
+    docker rm timescaledb
+fi
